requirements.txt
djangorestframework
djangorestframework-simplejwt
pugsql
